requirements.txt
## Hit By Pitches requirements.txt
## Author: Hossein Fuller <hossfuller@protonmail.com>
## Version: 1.0.0

## General operation
atproto
beautifulsoup4
colorama
ffmpeg-python
lxml
matplotlib
mysql-connector-python
numpy
pandas
pybaseball
Requests
scipy
tqdm
Unidecode

## Debugging, Documentation, etc.
build
black
mypy
no_implicit_optional
pipreqs
pytest
sphinx